requirements.txt
celery==3.1.17
Flask==0.10.1
flask_sqlalchemy==2.0
flask_wtf==0.11
Flask-Login==0.2.11
